# docs/api/gateway-openapi.yaml
openapi: 3.1.0
info:
  title: API Gateway
  version: 0.1.0
paths:
  /api/v1/health:
    get:
      summary: Gateway health check
      responses:
        '200':
          description: OK
  /api/v1/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200': { description: Authenticated }
  /api/v1/orders:
    post:
      summary: Submit order
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '202': { description: Order accepted for processing }
    get:
      summary: List orders
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Order list }
  /api/v1/portfolio/{accountId}:
    get:
      summary: Get account portfolio snapshot
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: accountId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Portfolio snapshot }
  /api/v1/strategies/backtests:
    post:
      summary: Start strategy backtest
      security: [{ bearerAuth: [] }]
      responses:
        '202': { description: Backtest accepted }
  /api/v1/copilot/query:
    post:
      summary: Ask AI copilot
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Copilot response }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateOrderRequest:
      type: object
      required: [accountId, instrumentId, side, orderType, quantity]
      properties:
        accountId: { type: string, format: uuid }
        instrumentId: { type: string, example: BTC-USD }
        side: { type: string, enum: [BUY, SELL] }
        orderType: { type: string, enum: [MARKET, LIMIT, STOP, STOP_LIMIT] }
        quantity: { type: number }
        price: { type: number }